feat(cli): add train command
Create speculators.train.cli with the training entrypoint (moved from
scripts/train.py), add speculators.train.__main__ for
`torchrun -m speculators.train`, and wire a thin typer dispatcher as
`speculators train` under the Pipeline help panel. Original script
becomes a deprecation shim. Update benchmark.py and test imports to
use new module paths.

Part of vllm-project/speculators#927 (Sub-PR 6 of 6)
