+++ src/speculators/cli/__init__.py
@@ -15,6 +15,7 @@ from speculators.cli.generate_data import generate_data
 from speculators.cli.prepare_data import prepare_data
 from speculators.cli.regenerate_responses import regenerate_responses
 from speculators.cli.stitch_mtp import stitch_command
+from speculators.cli.train import train_command
 
 __all__ = ["app"]
 
@@ -47,6 +48,11 @@ def _main(
 
 app.command(rich_help_panel="Pipeline")(prepare_data)
 app.command(rich_help_panel="Pipeline")(generate_data)
+app.command(
+    name="train",
+    rich_help_panel="Pipeline",
+    context_settings={"allow_extra_args": True, "allow_interspersed_args": False},
+)(train_command)
 app.command(rich_help_panel="Pipeline")(regenerate_responses)
 app.command(name="stitch-mtp", rich_help_panel="Pipeline")(stitch_command)
 app.command(rich_help_panel="Tools")(convert)
