+++ tests/unit/cli/test_cli.py
@@ -28,6 +28,7 @@ class TestRootApp:
         assert result.exit_code == 0
         assert "prepare-data" in result.output
         assert "generate-data" in result.output
+        assert "train" in result.output
         assert "regenerate-responses" in result.output
         assert "stitch-mtp" in result.output
 
@@ -116,16 +117,15 @@ class TestRegenerateResponsesCommand:
         assert result.exit_code != 0
 
 
-class TestGenerateDataCommand:
+class TestTrainCommand:
     def test_help(self):
-        result = runner.invoke(app, ["generate-data", "--help"])
+        result = runner.invoke(app, ["train", "--help"])
         assert result.exit_code == 0
-        assert "--endpoint" in result.output
-        assert "--preprocessed-data" in result.output
-        assert "--concurrency" in result.output
+        assert "--verifier-name-or-path" in result.output
+        assert "--config" in result.output
+        assert "--speculator-type" in result.output
 
-    def test_invalid_rank(self):
-        result = runner.invoke(
-            app, ["generate-data", "--rank", "5", "--world-size", "2"]
-        )
-        assert result.exit_code != 0
+    def test_train_appears_in_pipeline_panel(self):
+        result = runner.invoke(app, ["--help"])
+        assert result.exit_code == 0
+        assert "train" in result.output
